gpu-share/requirements.txt
# Core dependencies
numpy>=1.21.0
psutil>=5.8.0

# GPU libraries
pycuda>=2021.1
pyopencl>=2021.1.4
cupy>=10.0.0

# Communication
grpcio>=1.42.0
grpcio-tools>=1.42.0
websockets>=10.0
requests>=2.26.0
protobuf>=3.19.0

# Web framework
Flask>=2.0.2
Flask-CORS>=3.0.10
Jinja2>=3.0.3
Werkzeug>=2.0.2

# Utilities
pyyaml>=6.0
click>=8.0.3
python-dotenv>=0.19.2

# Testing
pytest>=6.2.5
pytest-cov>=3.0.0

# Logging
loguru>=0.5.3

# Security
cryptography>=36.0.0
PyJWT>=2.3.0

# Async support
aiohttp>=3.8.0
asyncio-mqtt>=0.11.0